major cleanup: letting go of rhst and old run_workflow!! [skip ci]
