--- neuropredict/__main__.py
@@ -1,15 +1,0 @@
-from sys import version_info
-
-if version_info.major > 2:
-    from neuropredict import run_workflow
-else:
-    raise NotImplementedError('neuropredict requires Python 3 or higher. '
-                              'Upgrade to Python 3+ or use [virtual] environments.')
-
-def main():
-    "Entry point."
-
-    run_workflow.cli()
-
-if __name__ == '__main__':
-    main()
