+++ neuropredict/reports.py
@@ -1,14 +1,15 @@
 
 import os
-from collections import Counter
-from os.path import join as pjoin, exists as pexists
 import pickle
-import numpy as np
+import shutil
 import traceback
+from collections import Counter
+from os.path import exists as pexists, join as pjoin
 
-from neuropredict import config_neuropredict as cfg
-from neuropredict import visualize, rhst
-from neuropredict.utils import load_options
+import numpy as np
+from neuropredict import config_neuropredict as cfg, visualize
+from neuropredict.utils import load_options, sub_group_identifier
+
 
 def export_results(dict_to_save, out_dir, options_path):
     """
@@ -180,12 +181,128 @@ def export_results_from_disk(results_file_path, out_dir, options_path):
     pred_prob_per_class, pred_labels_per_rep_fs, test_labels_per_rep, best_params,\
     feature_importances_rf, feature_names, num_times_misclfd, num_times_tested, \
     confusion_matrix, class_order, class_sizes, accuracy_balanced, auc_weighted, \
-    positive_class = classifier_name, feat_select_method = rhst.load_results(
+    positive_class = classifier_name, feat_select_method = load_results(
         results_file_path)
 
     locals_var_dict = locals()
     dict_to_save = {var: locals_var_dict[var]
                     for var in cfg.rhst_data_variables_to_persist}
     export_results(dict_to_save, out_dir, options_path)
+
+    return
+
+
+def save_results(out_dir, dict_of_objects_to_save):
+    "Serializes the results to disk."
+
+    # LATER choose a more universal serialization method
+    #   (that could be loaded from a web app)
+    try:
+        out_results_path = pjoin(out_dir, cfg.file_name_results)
+        with open(out_results_path, 'wb') as resfid:
+            pickle.dump(dict_of_objects_to_save, resfid)
+    except:
+        raise IOError('Error saving the results to disk!')
+    else:
+        # deleting temp results only when saving full results is successful
+        cleanup(out_dir)
+
+    return out_results_path
+
+
+def load_results_from_folder(results_folder):
+    """
+
+    Given a base output folder, possibly containing results for multiple sub-groups,
+        returns a dictionary of results, keyed in by sub group identifier.
+
+    """
+
+    results = dict()
+    options = load_options(results_folder)
+    for ix, sg in enumerate(options['sub_groups']):
+        sg_id = sub_group_identifier(sg, ix)
+        results_file_path = pjoin(results_folder, sg_id, cfg.file_name_results)
+        if not pexists(results_file_path) or os.path.getsize(results_file_path) <= 0:
+            raise IOError('Results file for sub group {} does not exist'
+                          ' or is empty!'.format(sg_id))
+        results[sg_id] = load_results_dict(results_file_path)
+
+    return results
+
+
+def load_results_dict(results_file_path):
+    "Loads the results serialized by RHsT."
+    # TODO need to standardize what needs to saved/read back
+
+    if not pexists(results_file_path) or os.path.getsize(results_file_path) <= 0:
+        raise IOError("Results file to be loaded doesn't exist, or empty!")
+
+    try:
+        with open(results_file_path, 'rb') as rf:
+            results_dict = pickle.load(rf)
+    except:
+        raise IOError('Error loading the saved results from \n{}'
+                      ''.format(results_file_path))
+
+    return results_dict
+
+
+def load_results(results_file_path):
+    "Loads the results serialized by RHsT."
+    # TODO need to standardize what needs to saved/read back
+
+    if not pexists(results_file_path):
+        raise IOError("Results file to be loaded doesn't exist!")
+
+    try:
+        with open(results_file_path, 'rb') as rf:
+            results_dict = pickle.load(rf)
+            # # below is possible, but not explicit and a bad practice
+            # # importing the keys and their values into the workspace
+            # locals().update(results_dict)
+
+            dataset_paths, method_names, train_perc, num_repetitions, num_classes, \
+            pred_prob_per_class, pred_labels_per_rep_fs, test_labels_per_rep, \
+            best_params, feature_importances_rf, \
+            feature_names, num_times_misclfd, num_times_tested, \
+            confusion_matrix, class_set, target_sizes, accuracy_balanced, \
+            auc_weighted, positive_class, classifier_name, feat_select_method = \
+                [results_dict.get(var_name)
+                 for var_name in cfg.rhst_data_variables_to_persist]
+
+    except:
+        raise IOError('Error loading the saved results from \n{}'
+                      ''.format(results_file_path))
+
+    # TODO need a consolidated way to identify the variables saved their order
+    return dataset_paths, method_names, train_perc, num_repetitions, num_classes, \
+           pred_prob_per_class, pred_labels_per_rep_fs, test_labels_per_rep, \
+           best_params, feature_importances_rf, feature_names, \
+           num_times_misclfd, num_times_tested, \
+           confusion_matrix, class_set, target_sizes, \
+           accuracy_balanced, auc_weighted, positive_class, \
+           classifier_name, feat_select_method
+
+
+def get_temp_dir(out_results_dir):
+    "Scratch directory to save temporary results to"
+
+    tmp_dir = pjoin(out_results_dir, cfg.temp_results_dir)
+    os.makedirs(tmp_dir, exist_ok=True)
+
+    return tmp_dir
+
+
+def cleanup(out_dir):
+    "Helper to perform cleanup"
+
+    tmp_dir = get_temp_dir(out_dir)
+    try:
+        shutil.rmtree(tmp_dir)
+    except:
+        traceback.print_exc()
+        print('Unable to delete temporary folder at:\n\t{}\n'
+              'Remove it manually if you would like to save space.'.format(tmp_dir))
 
     return
